// src/components/app/types.ts
import { LucideIcon } from 'lucide-react'

export type User = {
    name: string
    email: string
    avatar: string
}

export type BaseNavItem = {
    title: string
    badge?: string
    icon?: LucideIcon
    disabled?: boolean
}

export type NavLink = BaseNavItem & {
    url: string
    items?: never
}

export type NavCollapsible = BaseNavItem & {
    items: (BaseNavItem & { url: string })[]
    url?: never
}

export type NavItem = NavCollapsible | NavLink

export type NavGroup = {
    title: string
    items: NavItem[]
}

export type SidebarData = {
    user: User
    navGroups: NavGroup[]
}